requirements.txt
Django==3.2.1
djangorestframework==3.13.1
protobuf==5.26.1
